Move tinymultiindex
