+++ examples/harmonic_2D.cpp
@@ -2,7 +2,7 @@
 #include <fstream>
 
 #include "waveblocks/types.hpp"
-#include "waveblocks/tiny_multi_index.hpp"
+#include "waveblocks/wavepackets/shapes/tiny_multi_index.hpp"
 #include "waveblocks/matrixPotentials/potentials.hpp"
 #include "waveblocks/matrixPotentials/bases.hpp"
 #include "waveblocks/wavepackets/hawp_paramset.hpp"
@@ -33,7 +33,7 @@ int main() {
 
     const real_t eps = 0.1;
 
-    using MultiIndex = TinyMultiIndex<unsigned long, D>;
+    using MultiIndex = wavepackets::shapes::TinyMultiIndex<unsigned long, D>;
 
     // The parameter set of the initial wavepacket
     CMatrix<D,D> Q = CMatrix<D,D>::Identity();
