+++ examples/nonsense_2N_1D.cpp
@@ -5,7 +5,7 @@
 #include "waveblocks/propagators/Hagedorn.hpp"
 #include "waveblocks/matrixPotentials/potentials.hpp"
 #include "waveblocks/matrixPotentials/bases.hpp"
-#include "waveblocks/tiny_multi_index.hpp"
+#include "waveblocks/wavepackets/shapes/tiny_multi_index.hpp"
 #include "waveblocks/wavepackets/hawp_commons.hpp"
 #include "waveblocks/wavepackets/hawp_paramset.hpp"
 #include "waveblocks/wavepackets/shapes/shape_enumerator.hpp"
@@ -28,7 +28,7 @@ int main() {
 
     const real_t eps = 0.1;
 
-    using MultiIndex = TinyMultiIndex<unsigned short, D>;
+    using MultiIndex = wavepackets::shapes::TinyMultiIndex<unsigned short, D>;
 
     // The parameter set of the initial wavepacket
     CMatrix<D,D> Q = CMatrix<D,D>::Identity();
